requirements.txt
discord.py==1.5.1
APScheduler==0.0.0
discord==1.0.1
python-dotenv==0.15.0